+++ src/embeddingLib/src/embeddingSpace/Poincare.cpp
@@ -20,12 +20,12 @@ float Poincare::getSimilarity(NodeId a, NodeId b) const {
     float eps = 1e-5;
 
     // Squared norms, clamped
-    float sqanorm = std::min(std::max(coordinates[a].sqNorm(), 0.0), 1.0 - eps);
-    float sqbnorm = std::min(std::max(coordinates[b].sqNorm(), 0.0), 1.0 - eps);
+    float sqanorm = std::min(std::max(coordinates[a].sqNorm(), 0.0f), 1.0f - eps);
+    float sqbnorm = std::min(std::max(coordinates[b].sqNorm(), 0.0f), 1.0f - eps);
     float sqdist = tmpVec.sqNorm();
 
     float x = (sqdist / ((1 - sqanorm) * (1 - sqbnorm))) * 2 + 1;
-    float z = std::sqrt(std::pow(x, 2) - 1);
+    float z = std::sqrtf(std::powf(x, 2) - 1);
     return std::log(x + z);
 }
